requirements.txt
streamlit==0.73.1
Pillow == 7.2.0
numpy==1.18.5
opencv-python==4.4.0.46
